controllers, models  and routes modified

+++ controllers/shopController.js
@@ -1,78 +1,53 @@
 const Shop = require('../models/shopModel');
 const fs = require('fs/promises');
 
-// Create a new shop with single shopLogo and multiple shopBanners
+// Helper function to delete files
+const deleteFiles = async (files) => {
+    return Promise.all(files.map(async (file) => {
+        try {
+            await fs.unlink(file);
+        } catch (err) {
+            console.error(`Failed to delete file ${file}: ${err.message}`);
+        }
+    }));
+};
+
+// Create a new shop
 const createShop = async (req, res) => {
-    let shopLogo;
-    let shopBanners;
+    const { name, description, ownerId } = req.body;
+    const shopLogo = req.files['shopLogo'] ? req.files['shopLogo'][0].path : null;
+    const shopBanners = req.files['shopBanners'] ? req.files['shopBanners'].map(file => file.path) : [];
+
     try {
-        const { name, description, ownerId } = req.body;
-
-        shopLogo = req.files['shopLogo'] ? req.files['shopLogo'][0].path : null;
-        shopBanners = req.files['shopBanners'] ? req.files['shopBanners'].map(file => file.path) : [];
-
-        console.log(name, description, ownerId, shopLogo, shopBanners);
-
-        const shop = new Shop({
-            name,
-            description,
-            ownerId,
-            shopLogo,
-            shopBanners
-        });
-
+        const shop = new Shop({ name, description, ownerId, shopLogo, shopBanners });
         await shop.save();
         res.status(201).json(shop);
     } catch (error) {
-        if (shopLogo) {
-            try {
-                await fs.unlink(shopLogo);
-                console.error(`Deleted orphaned image successfully: ${err.message}`);
-            } catch (err) {
-                console.error(`Failed to delete orphaned image: ${err.message}`);
-            }
-        }
-        if (shopBanners) {
-            try {
-                await fs.unlink(shopBanners);
-                console.error(`Deleted orphaned images successfully: ${err.message}`);
-            } catch (err) {
-                console.error(`Failed to delete orphaned images: ${err.message}`);
-            }
-        }
+        if (shopLogo) await fs.unlink(shopLogo).catch(err => console.error(`Failed to delete orphaned logo: ${err.message}`));
+        await deleteFiles(shopBanners);
         res.status(500).json({ message: 'Failed to create shop', error: error.message });
     }
 };
 
-// Update an existing shop (with single logo and multiple banners)
+// Update an existing shop
 const updateShop = async (req, res) => {
-    let newShopLogo;
-    let newShopBanners;
+    const { shopId } = req.params;
+    const { name, description } = req.body;
+    const shop = await Shop.findById(shopId);
+
+    if (!shop) return res.status(404).json({ message: 'Shop not found' });
+
     try {
-        const { shopId } = req.params;
-        const { name, description } = req.body;
-
-        const shop = await Shop.findById(shopId);
-        if (!shop) return res.status(404).json({ message: 'Shop not found' });
-
-        // Handle new logo upload and delete the old one if it exists
+        // Handle logo update
         if (req.files['shopLogo']) {
-            if (shop.shopLogo) {
-                await fs.unlink(shop.shopLogo).catch((err) => console.log('Failed to delete old logo', err));
-            }
-            newShopLogo = req.files['shopLogo'][0].path;
-            shop.shopLogo = newShopLogo;
+            if (shop.shopLogo) await fs.unlink(shop.shopLogo).catch(err => console.log('Failed to delete old logo', err));
+            shop.shopLogo = req.files['shopLogo'][0].path;
         }
 
-        // Handle new banners upload and delete the old ones if they exist
+        // Handle banners update
         if (req.files['shopBanners']) {
-            if (shop.shopBanners.length > 0) {
-                await Promise.all(shop.shopBanners.map(async (banner) => {
-                    await fs.unlink(banner).catch((err) => console.log('Failed to delete old banner', err));
-                }));
-            }
-            newShopBanners = req.files['shopBanners'].map(file => file.path);
-            shop.shopBanners = newShopBanners;
+            await deleteFiles(shop.shopBanners);
+            shop.shopBanners = req.files['shopBanners'].map(file => file.path);
         }
 
         // Update name and description
@@ -82,50 +57,30 @@ const updateShop = async (req, res) => {
         await shop.save();
         res.status(200).json(shop);
     } catch (error) {
-        if (newShopLogo) {
-            try {
-                await fs.unlink(newShopLogo);
-                console.error(`Deleted orphaned image successfully: ${err.message}`);
-            } catch (err) {
-                console.error(`Failed to delete orphaned image: ${err.message}`);
-            }
-        }
-        if (newShopBanners) {
-            try {
-                await fs.unlink(newShopBanners);
-                console.error(`Deleted orphaned images successfully: ${err.message}`);
-            } catch (err) {
-                console.error(`Failed to delete orphaned images: ${err.message}`);
-            }
-        }
+        if (shop.shopLogo) await fs.unlink(shop.shopLogo).catch(err => console.error(`Failed to delete orphaned logo: ${err.message}`));
+        await deleteFiles(req.files['shopBanners'] ? req.files['shopBanners'].map(file => file.path) : []);
         res.status(500).json({ message: 'Failed to update shop', error: error.message });
     }
 };
 
-// Get a single shop by ID
+// Get a single shop
 const getSingleShop = async (req, res) => {
     try {
         const { shopId } = req.params;
         const shop = await Shop.findById(shopId);
-
         if (!shop) return res.status(404).json({ message: 'Shop not found' });
-
         res.status(200).json(shop);
     } catch (error) {
         res.status(500).json({ message: 'Failed to retrieve shop', error: error.message });
     }
 };
 
-// Get all shops for a specific user
+// Get all shops for a specific Merchant
 const getMerchantShops = async (req, res) => {
     try {
         const { merchantId } = req.params;
         const shops = await Shop.find({ ownerId: merchantId });
-
-        if (shops.length === 0) {
-            return res.status(404).json({ message: 'No shops found for this user' });
-        }
-
+        if (shops.length === 0) return res.status(404).json({ message: 'No shops found for this user' });
         res.status(200).json(shops);
     } catch (error) {
         res.status(500).json({ message: 'Failed to retrieve shops for user', error: error.message });
@@ -142,27 +97,16 @@ const getAllShops = async (req, res) => {
     }
 };
 
-// Delete a shop (with promise-based image deletion)
+// Delete a shop
 const deleteShop = async (req, res) => {
-    try {
-        const { shopId } = req.params;
-        console.log(shopId);
+    const { shopId } = req.params;
 
+    try {
         const shop = await Shop.findById(shopId);
         if (!shop) return res.status(404).json({ message: 'Shop not found' });
 
-        console.log("shop found", shop);
-
-        // Delete logo and banners if they exist
-        if (shop.shopLogo) {
-            await fs.unlink(shop.shopLogo).catch((err) => console.log('Failed to delete shop logo', err));
-        }
-
-        if (shop.shopBanners.length > 0) {
-            await Promise.all(shop.shopBanners.map(async (banner) => {
-                await fs.unlink(banner).catch((err) => console.log('Failed to delete shop banner', err));
-            }));
-        }
+        // Delete logo and banners
+        await deleteFiles([shop.shopLogo, ...shop.shopBanners]);
 
         await Shop.findByIdAndDelete(shopId);
         res.status(200).json({ message: 'Shop deleted successfully' });
@@ -172,3 +116,4 @@ const deleteShop = async (req, res) => {
 };
 
 module.exports = { createShop, updateShop, getSingleShop, getMerchantShops, getAllShops, deleteShop };
+
